pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[project]
name = "stephen"
authors = [
    {name = "Antmicro", email="contact@antmicro.com"}
]
license = "Apache-2.0"

description = "STEPhen - STEP output job automation tool"
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "cadquery", "python-slugify", "pandas", "typer", "GitPython", "Jinja2", "pandas-stubs", "types-python-slugify"
]

[project.scripts]
stephen = "stephen.main:app"

[tool.ruff]
select = ["E", "F", "N", "B", "D", "A", "RET"]
line-length = 120
target-version = "py311"
exclude = [
	"documentation",
	"__init__.py",
]

[tool.black]
line-length = 120

[tool.mypy]
disallow_any_generics = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true